+++ src/qgis_geonode/network.py
@@ -277,237 +277,6 @@ class NetworkRequestTask(qgis.core.QgsTask):
         log("inside _handle_auth_browser_aborted")
 
 
-class LayerUploaderTask(NetworkRequestTask):
-    VECTOR_UPLOAD_FORMAT: typing.Final[str] = "gpkg"
-    RASTER_UPLOAD_FORMAT: typing.Final[str] = "tif"
-
-    layer: qgis.core.QgsMapLayer
-    allow_public_access: bool
-    _upload_url: QtCore.QUrl
-    _temporary_directory: typing.Optional[Path]
-
-    def __init__(
-        self,
-        layer: qgis.core.QgsMapLayer,
-        upload_url: QtCore.QUrl,
-        allow_public_access: bool,
-        authcfg: str,
-        network_task_timeout: int,
-        description: str = "LayerUploaderTask",
-    ):
-        super().__init__(
-            requests_to_perform=[],
-            authcfg=authcfg,
-            description=description,
-            network_task_timeout=network_task_timeout,
-        )
-        self.response_contents = [None]
-        self.layer = layer
-        self.allow_public_access = allow_public_access
-        self._upload_url = upload_url
-        self._temporary_directory = None
-
-    def run(self) -> bool:
-        if self._is_layer_uploadable():
-            log("Layer is in a format natively supported, no need to export.")
-            source_path = self.layer.dataProvider().dataSourceUri().partition("|")[0]
-            export_error = None
-        else:  # we need to export the layer first
-            log(
-                "Exporting layer to an uploadable format before proceeding with "
-                "the upload..."
-            )
-            source_path, export_error = self._export_layer_to_temp_dir()
-        if export_error is None:
-            log(f"exported data is in {source_path}")
-            # TODO: check if source path is an actual path
-            source_file = QtCore.QFile(str(source_path))
-            source_file.open(QtCore.QIODevice.ReadOnly)
-            payload = self._build_multipart(source_file, Path(source_path).name)
-            source_file.setParent(payload)
-            with wait_for_signal(
-                self._all_requests_finished, timeout=self.network_task_timeout
-            ) as event_loop_result:
-                request = QtNetwork.QNetworkRequest(self._upload_url)
-                request.setHeader(
-                    QtNetwork.QNetworkRequest.ContentTypeHeader,
-                    f"multipart/form-data; boundary={payload.boundary().data().decode()}",
-                )
-                if self.authcfg:
-                    auth_manager = qgis.core.QgsApplication.authManager()
-                    auth_added, _ = auth_manager.updateNetworkRequest(
-                        request, self.authcfg
-                    )
-                else:
-                    auth_added = True
-                if auth_added:
-                    qt_reply = self._dispatch_request(request, HttpMethod.POST, payload)
-                    payload.setParent(qt_reply)
-                    request_id = qt_reply.property("requestId")
-                    self._pending_replies[request_id] = (0, qt_reply)
-                else:
-                    self._all_requests_finished.emit()
-            loop_forcibly_ended = not bool(event_loop_result.result)
-            log(f"loop_forcibly_ended: {loop_forcibly_ended}")
-            if loop_forcibly_ended:
-                result = False
-            else:
-                result = self._num_finished >= len(self.requests_to_perform)
-        else:
-            result = False
-        return result
-
-    def finished(self, result: bool) -> None:
-        if self._temporary_directory is not None:
-            log(
-                f"About to delete the temporary directory "
-                f"at {self._temporary_directory} ..."
-            )
-            # shutil.rmtree(self._temporary_directory, ignore_errors=True)
-        super().finished(result)
-
-    def _is_layer_uploadable(self) -> bool:
-        ds_uri = self.layer.dataProvider().dataSourceUri()
-        fragment = ds_uri.split("|")[0]
-        extension = fragment.rpartition(".")[-1]
-        return extension in (self.VECTOR_UPLOAD_FORMAT, self.RASTER_UPLOAD_FORMAT)
-
-    def _export_layer_to_temp_dir(
-        self,
-    ) -> typing.Tuple[typing.Optional[Path], typing.Optional[str]]:
-        self._temporary_directory = Path(tempfile.mkdtemp(prefix="qgis_geonode_"))
-        log(f"inside _export_layer_to_temp_dir")
-        exported_path = None
-        error_message = None
-        if self.layer.type() == qgis.core.QgsMapLayerType.VectorLayer:
-            log("is vector")
-            exported_path, error_message = self._export_vector_layer()
-        elif self.layer.type() == qgis.core.QgsMapLayerType.RasterLayer:
-            log("is raster")
-            exported_path, export_error = self._export_raster_layer()
-        else:
-            log("is unknown - panic!")
-            raise NotImplementedError()
-        return exported_path, error_message or None
-
-    def _export_vector_layer(self) -> typing.Tuple[typing.Optional[Path], str]:
-        sanitized_layer_name = sanitize_layer_name(self.layer.name())
-        log("inside _export_vector_layer")
-        target_path = (
-            self._temporary_directory
-            / f"{sanitized_layer_name}.{self.VECTOR_UPLOAD_FORMAT}"
-        )
-        log(f"target_path: {target_path}")
-        export_code, error_message = qgis.core.QgsVectorLayerExporter.exportLayer(
-            layer=self.layer,
-            uri=str(target_path),
-            providerKey="ogr",
-            destCRS=qgis.core.QgsCoordinateReferenceSystem(),
-            onlySelected=True,
-            options={
-                "driverName": self.VECTOR_UPLOAD_FORMAT.upper(),
-                "layerName": sanitized_layer_name,
-            },
-        )
-        log(f"export_code: {export_code}")
-        log(f"error_message: {error_message}")
-        if export_code == qgis.core.Qgis.VectorExportResult.Success:
-            result = (target_path, error_message)
-        else:
-            result = (None, error_message)
-        log("leaving _export_vector_layer...")
-        return result
-
-    def _export_raster_layer(
-        self,
-    ) -> typing.Tuple[typing.Optional[Path], typing.Optional[int]]:
-        sanitized_layer_name = sanitize_layer_name(self.layer.name())
-        target_path = (
-            self._temporary_directory
-            / f"{sanitized_layer_name}.{self.RASTER_UPLOAD_FORMAT}"
-        )
-        writer = qgis.core.QgsRasterFileWriter(str(target_path))
-        writer.setOutputFormat("GTiff")
-        pipe = self.layer.pipe()
-        raster_interface = self.layer.dataProvider()
-        write_error = writer.writeRaster(
-            pipe,
-            raster_interface.xSize(),
-            raster_interface.ySize(),
-            raster_interface.extent(),
-            raster_interface.crs(),
-            qgis.core.QgsCoordinateTransformContext(),
-        )
-        if write_error == qgis.core.QgsRasterFileWriter.NoError:
-            result = (target_path, None)
-        else:
-            result = (None, write_error)
-        return result
-
-    def _build_multipart(
-        self, dataset_contents: QtCore.QIODevice, dataset_filename: str
-    ) -> QtNetwork.QHttpMultiPart:
-        encoding = "utf-8"
-        multipart = QtNetwork.QHttpMultiPart(QtNetwork.QHttpMultiPart.FormDataType)
-        metadata = self.layer.metadata()
-        title_part = QtNetwork.QHttpPart()
-        title_part.setHeader(
-            QtNetwork.QNetworkRequest.ContentDispositionHeader,
-            'form-data; name="dataset_title"',
-        )
-        title_part.setBody(metadata.title().encode(encoding))
-        multipart.append(title_part)
-        abstract_part = QtNetwork.QHttpPart()
-        abstract_part.setHeader(
-            QtNetwork.QNetworkRequest.ContentDispositionHeader,
-            'form-data; name="abstract"',
-        )
-        abstract_part.setBody(metadata.abstract().encode(encoding))
-        multipart.append(abstract_part)
-        false_items = (
-            "time",
-            "mosaic",
-            "metadata_uploaded_preserve",
-            "metadata_upload_form",
-            "style_upload_form",
-        )
-        for item in false_items:
-            part = QtNetwork.QHttpPart()
-            part.setHeader(
-                QtNetwork.QNetworkRequest.ContentDispositionHeader,
-                f'form-data; name="{item}"',
-            )
-            part.setBody("false".encode("utf-8"))
-            multipart.append(part)
-        permissions_part = QtNetwork.QHttpPart()
-        permissions_part.setHeader(
-            QtNetwork.QNetworkRequest.ContentDispositionHeader,
-            'form-data; name="permissions"',
-        )
-        permissions = {
-            "users": {},
-            "groups": {},
-        }
-        if self.allow_public_access:
-            permissions["users"]["AnonymousUser"] = [
-                "view_resourcebase",
-                "download_resourcebase",
-            ]
-        permissions_part.setBody(json.dumps(permissions).encode(encoding))
-        multipart.append(permissions_part)
-        file_part = QtNetwork.QHttpPart()
-        file_part.setHeader(
-            QtNetwork.QNetworkRequest.ContentDispositionHeader,
-            f'form-data; name="base_file"; filename="{dataset_filename}"',
-        )
-        file_part.setHeader(
-            QtNetwork.QNetworkRequest.ContentTypeHeader, "application/x-qgis"
-        )
-        file_part.setBodyDevice(dataset_contents)
-        multipart.append(file_part)
-        return multipart
-
-
 def deserialize_json_response(
     contents: QtCore.QByteArray,
 ) -> typing.Optional[typing.Union[typing.List, typing.Dict]]:
